test
test

+++ fonksiyon.c
@@ -11,9 +11,10 @@ int cikar(int sayi1, int sayi2) {
 }
 
 // Çarpma Fonksiyonu
-int carp(int sayi1, int sayi2) {
+float carp(float sayi1, float sayi2) {
     return sayi1 * sayi2;
 }
+
 
 // Bölme Fonksiyonu
 float bol(float sayi1, float sayi2) {
@@ -38,8 +39,8 @@ int main() {
     printf("%d - %d = %d\n", sayi1, sayi2, sonuc2);
 
     // Çarpma işlemi
-    int sonuc3 = carp(sayi1, sayi2);
-    printf("%d * %d = %d\n", sayi1, sayi2, sonuc3);
+    float sonuc3 = carp(sayi1, sayi2);
+    printf("%d * %d = %.2f\n", sayi1, sayi2, sonuc3);
 
     // Bölme işlemi
     float sonuc4 = bol(sayi1, sayi2);
